pyproject.toml
[project]
name = "magic-marketplace"
version = "0.1.0"
description = ""
authors = [
    {name = "Guilherme Iazzetta",email = "guilherme.iazzetta@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.119.1,<0.120.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "aerich[toml] (>=0.9.2,<0.10.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-order (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "src.configuration.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
